+++ examples/nuxt/entity-queries/nuxt.config.js
@@ -1,7 +1,11 @@
+const baseUrl = process.env.GITPOD_WORKSPACE_ID
+  ? `https://8080-${process.env.GITPOD_WORKSPACE_ID}.${process.env.GITPOD_WORKSPACE_CLUSTER_HOST}`
+  : process.env.BASE_URL || 'http://drupal-9.ddev.site'
+
 export default {
   buildModules: ['druxt-entity'],
   druxt: {
-    baseUrl: 'https://demo-api.druxtjs.org',
+    baseUrl,
     entity: {
       query: {
         // Enable auto-filtering fields by schema: