+++ examples/nuxt/menu-edit/nuxt.config.js
@@ -1,3 +1,7 @@
+const baseUrl = process.env.GITPOD_WORKSPACE_ID
+  ? `https://8080-${process.env.GITPOD_WORKSPACE_ID}.${process.env.GITPOD_WORKSPACE_CLUSTER_HOST}`
+  : process.env.BASE_URL || 'http://drupal-9.ddev.site'
+
 export default {
   components: true,
   modules: [
@@ -5,7 +9,7 @@ export default {
     'druxt-menu',
   ],
   druxt: {
-    baseUrl: 'https://demo-api.druxtjs.org/',
+    baseUrl,
     menu: {
       jsonApiMenuItems: true,
     },