+++ examples/nuxt/schema-component/nuxt.config.js
@@ -1,4 +1,8 @@
+const baseUrl = process.env.GITPOD_WORKSPACE_ID
+  ? `https://8080-${process.env.GITPOD_WORKSPACE_ID}.${process.env.GITPOD_WORKSPACE_CLUSTER_HOST}`
+  : process.env.BASE_URL || 'http://drupal-9.ddev.site'
+
 export default {
-  buildModules: [['druxt-schema', { baseUrl: 'https://demo-api.druxtjs.org' }]],
+  buildModules: [['druxt-schema', { baseUrl }]],
   components: true
 }