+++ packages/views/CHANGELOG.md
@@ -1,5 +1,16 @@
 # druxt-views
 
+## 0.18.2
+
+### Patch Changes
+
+- c6d7c83: Fixed available component options.
+- 540afca: Updated drupal-jsonapi-params to 2.0.0
+- Updated dependencies [540afca]
+  - druxt-blocks@0.15.2
+  - druxt@0.18.2
+  - druxt-entity@0.24.3
+
 ## 0.18.1
 
 ### Patch Changes